requirements.txt
# snowplow-tracker
# requests
# typing-extensions

certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
requests==2.32.3
snowplow-tracker==1.0.3
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
pyspark==3.5.3
python-dotenv==1.0.1